Update Turkish Transloter
When translating a language, it should be done with the translation of its own Language Association, not with Google sites or Google Translate. According to the Turkish Language Association, these words are written incorrectly. It has been corrected in the current file.I hope Greg notices my efforts and provides a "translator".Kind regards..

Signed-off-by: Savas400-star <zekeriyaozturk428@gmail.com>