Update Arabic ar Language (#366)
Line 330: The word "العلم" was changed to "العَلَمِ" for clearer meaning, as "العلم" has two meanings.

Line 332: The preposition "في" was added because it is more commonly used.

Line 405: A space was added to separate the number from the word.

Line 441: The preposition "من" was added instead of "عن" because "من" is more appropriate, common, and correct in this sentence's context.

Line 550: The sentence was rephrased for smoother and more accurate expression, correcting the error of mixing members and friends.

Line 791: A simple and direct style was adopted.

Line 794: A spelling mistake was corrected.

Line 882: A space was added between the two words.

Line 997: The letter "ه" was replaced with "ة" as it is the correct form.

Line 1118: A translation error was corrected here.

Line 1131: The entire sentence was changed to make it easier to understand; even I had difficulty understanding it when I read it.

Line 1136: The modification was made; "لون" should precede the item, such as "لون القلم" or "لون الاسم". The correct sentence is "لون المستوى"; "لون" precedes the item. "In Arabic Language"

Line 1145: The omitted letter was added.

Line 1249: The incorrect sentence was revised.

Signed-off-by: MaRs001V <mahamed.med2005@gmail.com>