Change some of the words of the German language
Hello, I have modified some German words to be more common for the player who reads the first change is the word
In line 237 I changed the word (Masse) and replaced it with (Menge) because I found it more common. 
2 In line 133 I changed the word Global and replaced it with Weltweit because I found it more popular and aesthetically pleasing.
In line 131 I changed the word Fehler to Fehler because I found it more common and because it is a word that is frequently used. Problem because I find it more common and because it is a word that is used a lot
In line 113 I changed the word Warnung to Hinweis because I find it beautiful and many people use it 
In line 112 I changed the word Freunde to Freundesliste because it expresses the place where friends are found and they can be unfriends because I add them to take a temporary favor from them and they are considered unfriends.

Signed-off-by: ahmdhjknj <z50920386@gmail.com>