Indonesia ( IN ) Improvements Translation
My ID : 14001294

The reason I made the translation improvements was so that the language used in this game would be easier to understand for people of all ages, Many of the languages used previously were a little less easy to understand, so since the majority of nebulous game players are mostly under 20 years old, Or you could say that many of the users are teenagers and adults, so I took the initiative to create a language that is easy enough to recognize and understand for all age groups. That's all and thank you!

Signed-off-by: 7Shann <bapakhaji156@gmail.com>