🇬🇷 `el` Greek UI-related Improvements
Hi!
You might know me from contributing to the Polish language or commenting. A long time ago, I posted a suggestion regarding the length of texts, which are too long, hoping that Greek translators would come forward and the problem would be solved. Since there are still no such people, I decided to propose specific changes that might turn out to be better.
## Line 122, 123, 124 (Problem 1)
Probably the most important change in the entire PR, which shortens the drawn-out text (especially: Offline) from a vertical perspective. For now, I would leave it as it is ( Online, Άγνωστο, Offline, as it is the lesser evil. 
## Line 1245
This could be shortened for UI reasons, as you can see in the screenshot, the current version is also lengthy. The meaning of the proposed translation does not change either.
## Lines: 979, 980
Similar situation here, a long translation results in a very large UI, which obscures the field of view. I suggest shortening this, because this word does not particularly require the UI to be larger just because of it. Ζωάκι = little pet
## Lines: 910, 911, 912, 934, 936, 981, 1077, 1224, 1227, 1271
All follow Consistency in Terminology rule. The strings have been adapted to fit the following forms: Ζωάκια (pets), Ζωάκι (pet). They are case-sensitive: `ζ` - small `z` ; `Z` - capitalised `Z`, and finally: grammar sensitive.



Signed-off-by: Ania <136233479+AnkaKoxu@users.noreply.github.com>